pyproject.toml
[tool.poetry]
name = "volby"
version = "0.1.0"
description = ""
authors = ["Jan Kislinger <jan.kislinger@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.1"
numpy = "^1.23.4"
pyjanitor = "^0.24.0"
plotly = "^5.11.0"
pyarrow = "^10.0.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.7.7"
isort = "^5.10.1"
flake8 = "^5.0.4"
black = "^22.10.0"

[tool.black]
max-line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
